Stitch truncated log lines back together before display
When fetched via the API, logs are split into lines with a position and
timestamp. However, there's also an upper limit to log line length of
16k; lines longer than this will be split on the server.

When dealing with color output, this has the potential to split in the
middle of an ANSI color code, meaning that when displayed in
ansi-terminal, partial codes can appear in plain text at the beginning
and end of lines, and colors can fail to be reset correctly.

This patch attempts to address the issue by stitching together log lines
truncated in this manner. It works by inspecting the last character of
the last log line seen, and appending the current line's log output
unless the last line ended with a newline character.

Because log line objects are immutable, we have to clone the lines while
stitching; we also need to update the line numbers to account for log
lines that are joined together.