Truncate commit message to first line in RepoItem
RepoItems attempt to display the entire commit message for each build
triggered by a commit. This is okay for most users as the message will
run out of space and be truncated. However, I among many others with
Hi-DPI displays have customized Drone's CSS to take up more monitor
space, which leads to some very ugly build names, as the RepoItem
displays the commit message and body.

This PR trims the title to just the first line of the commit message.